Added information.
Finished the About Me section. Added the social media links. Changed
the Portfolio Subheadings. Took out the PALETTA project because it is
not done. Took out a lot of content in the ECP because I want to put in
more images and less words. The Images have to tell the story.